Small cosmetic cleanup.
